refactor: enhance product search and verification logic, improve toast component behavior

+++ backend/voiceverification/db/speaker_repo.py
@@ -44,7 +44,7 @@ def load_all_embeddings(user_id: str) -> list[np.ndarray]:
     profiles = []
     for row in res.data:
         profiles.append({
-            "embedding": np.array(row["embedding"]),
+            "embedding": np.array(row["embedding"], dtype=np.float32),
             "label": row["label"]
         })
